+++ addon/locales/en/components/fd-folder-editing-panel.js
@@ -1,0 +1,6 @@
+export default {
+  caption: 'Сaption',
+  'caption-placeholder': 'Enter caption',
+  'sub-form': 'Form',
+  'sub-folder': 'Subfolder',
+};