+++ addon/objects/uml-primitives/fd-uml-class.js
@@ -195,6 +195,8 @@ joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
 
 joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
 
+joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
+
 /**
   Defines the JointJS element, which represents the UML class in the diagram.
 
@@ -207,6 +209,8 @@ joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
 export let Class = BaseClass.define('flexberry.uml.Class', {});
 joint.util.setByPath(joint.shapes, 'flexberry.uml.Class', Class, '.');
 
+
+joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
 
 joint.util.setByPath(joint.shapes, 'flexberry.uml.BaseClass', BaseClass, '.');
 
@@ -313,19 +317,11 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }.bind(this));
 
     this.$box.find('.class-stereotype-input').on('blur', function(evt) {
-      let stereotypeText = $(evt.target).val();
-      let stereotype = this.normalizeStereotype(stereotypeText);
-      let rows = stereotypeText.split(/[\n\r|\r|\n]/);
-      let $stereotypeInput = this.$box.find('.class-stereotype-input');
-      $stereotypeInput.val(stereotype);
-      $stereotypeInput.prop('rows', rows.length);
-      let objectModel = this.model.get('objectModel');
-      objectModel.set('stereotype', stereotype);
-      this.paper.trigger('updaterepobj', objectModel, 'stereotype', stereotype);
-      this.updateRectangles();
+      this.showNormalizedStereotypeOnInput($(evt.target));
     }.bind(this));
 
     this.updateInputValue();
+    this.showNormalizedStereotypeOnInput(this.$box.find('.class-stereotype-input'));
 
     // Update the box position whenever the underlying model changes.
     this.model.on('change', this.updateBox, this);
@@ -504,5 +500,14 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }
 
     return stereotype;
+  },
+
+  showNormalizedStereotypeOnInput(element) {
+    let stereotypeText = element.val();
+    let stereotype = this.normalizeStereotype(stereotypeText);
+    let rows = stereotypeText.split(/[\n\r|\r|\n]/);
+    element.val(stereotype);
+    element.prop('rows', rows.length);
+    this.updateRectangles();
   }
 });