+++ app/components/fd-navigation-editing-panels/fd-create-node-editing-panel.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/components/fd-navigation-editing-panels/fd-create-node-editing-panel';