+++ tests/unit/routes/fd-system-list-form-test.js
@@ -4,11 +4,13 @@ moduleFor('route:fd-system-list-form', 'Unit | Route | fd system list form', {
   // Specify the other units that are required for this test.
   // needs: ['controller:foo']
   needs: [
+    'service:cols-config-menu',
     'service:fd-current-project-context',
     'service:objectlistviewEvents',
     'service:formLoadTimeTracker',
     'service:colsConfigMenu',
-    'service:appState'
+    'service:appState',
+    'service:adv-limit',
   ]
 });