+++ tests/unit/routes/fd-view-edit-form-test.js
@@ -4,7 +4,8 @@ moduleFor('route:fd-view-edit-form', 'Unit | Route | fd view edit form', {
   // Specify the other units that are required for this test.
   // needs: ['controller:foo']
   needs: [
-    'service:appState'
+    'service:appState',
+    'service:adv-limit',
   ]
 });